Move content script related files to their own folder
